feat(info-for-all-exercises): Show all infos for an exercise
This commit edits the wger/exercises/api/serializers to add  the exercise information
serializer to convert the model output into json serializable objects, wger/exercises/api/views.py
is also edited to add the model viewset for displaying the json objects. It adds modifications to
wger/urls.py to register the routes for the new readonly api-endpoint.

It adds tests for the api endpoint created

[Delivers #157731566]