feature(body-part-exercises-on-hover): Show body part exercises on hover
This commit enables a user to get all exercises for a body part when the mouse
hovers over the body part.

it adds new files svg.js which contains the svgs that the user hovers over
to view all the exercises associated with that body part.muscle-back.svg and
muscle-front.svg with all the muscles.

The files modified include are workout-manager.css to add the tranformation
css to fit the svg and hide it from the users. overview.html is edited to
add the function to get the svg paths on hovering and highlight them while
providing the necessary exercises.

[Delivers #157731584]